Moving functionality from prior commits to forcefield_utils.py rather than in a test module, as it's more generally useful.
