+++ examples/README.md
@@ -3,3 +3,4 @@
 ## Manifest
 * `parm@frosst/` - example illustrating attempt to recover parm@frosst atom types
 * `AlkEtOH/` - example illustrating attempt to recover parm99 atom types for a small set of alkanes, ethers, and alcohols
+* `SMIRFF_comparison/` - temporary example (waiting for permanent home) of cross-comparison of molecule energies from SMIRFF with same molecule energies from .prmtop and .crd.
